requirements.txt
gradio
diffusers
transformers
accelerate
peft
pillow
sdnq
torchvision
torchaudio
